refactor: move timers to OpState
