+++ cli/state.rs
@@ -19,7 +19,6 @@ use std::pin::Pin;
 use std::rc::Rc;
 use std::str;
 use std::sync::Arc;
-use std::time::Instant;
 
 // This is named "CliState" instead of just "State" to avoid confusion with all
 // other state structs (GlobalState, OpState, GothamState).
@@ -32,7 +31,6 @@ pub struct CliState {
   /// When flags contains a `.import_map_path` option, the content of the
   /// import map file will be resolved and set.
   pub import_map: Option<ImportMap>,
-  pub start_time: Instant,
   pub target_lib: TargetLib,
   pub is_main: bool,
   pub is_internal: bool,
@@ -155,7 +153,6 @@ impl CliState {
         .unwrap_or_else(|| global_state.permissions.clone())
         .into(),
       import_map: maybe_import_map,
-      start_time: Instant::now(),
       target_lib: TargetLib::Main,
       is_main: true,
       is_internal,
@@ -176,7 +173,6 @@ impl CliState {
         .unwrap_or_else(|| global_state.permissions.clone())
         .into(),
       import_map: None,
-      start_time: Instant::now(),
       target_lib: TargetLib::Worker,
       is_main: false,
       is_internal: false,
